+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ## Fixed
 
 -   Fixed faulty synchronization of chapter text when multiple clients are editing a book that led to temporarily missing chapter text.
+-   Fixed jumpy smooth scrolling of caret position.
 
 # 0.5.63 - 2023-11-11